+++ resources/js/types/index.d.ts
@@ -2,7 +2,8 @@ import { LucideIcon } from 'lucide-react';
 import type { Config } from 'ziggy-js';
 
 export interface Auth {
-    user: User;
+    user: User | null;
+    permissions: string[];
 }
 
 export interface BreadcrumbItem {